Feat: Add Ko-fi link and reusable footer
