+++ index.php
@@ -122,125 +122,5 @@
     </section>
 </main>
 
-<footer>
-    <div class="container">
-        <p>&copy; 2025 TunnelCraft</p>
-        <ul class="social-links">
-            <li><a href="#"><img src="assets/images/FB_Logo.png" alt="Facebook"></a></li>
-            <li><a href="#"><img src="assets/images/Linkedin_Logo.png" alt="LinkedIn"></a></li>
-            <li><a href="#"><img src="assets/images/GitHub_Logo.png" alt="GitHub"></a></li>
-        </ul>
-    </div>
-</footer>
-
-<script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-app-compat.js"></script>
-<script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-auth-compat.js"></script>
-
-<?php require_once 'config.php'; ?>
-
-<script>
-  const menuToggle = document.querySelector('.menu-toggle');
-    const navLinks = document.querySelector('.nav-links');
-
-    menuToggle.addEventListener('click', () => {
-        navLinks.classList.toggle('active');
-    });
-
-    // Carousel functionality
-    document.addEventListener('DOMContentLoaded', () => {
-    const carouselSlide = document.querySelector('.carousel-slide');
-    const prevBtn = document.querySelector('.prev-btn');
-    const nextBtn = document.querySelector('.next-btn');
-
-    let counter = 0;
-    const totalTiles = carouselSlide.querySelectorAll('.project-tile').length;
-
-    function updateCarousel() {
-        // Calculate the number of visible tiles
-        const visibleTiles = window.innerWidth > 768 ? 3 : 1;
-        
-        // Disable buttons at the start and end of the carousel
-        if (counter === 0) {
-            prevBtn.disabled = true;
-        } else {
-            prevBtn.disabled = false;
-        }
-
-        if (counter >= totalTiles - visibleTiles) {
-            nextBtn.disabled = true;
-        } else {
-            nextBtn.disabled = false;
-        }
-
-        const tileSize = carouselSlide.querySelector('.project-tile').clientWidth;
-        carouselSlide.style.transform = `translateX(${-tileSize * counter}px)`;
-    }
-
-    nextBtn.addEventListener('click', () => {
-        const visibleTiles = window.innerWidth > 768 ? 3 : 1;
-        if (counter < totalTiles - visibleTiles) {
-            counter++;
-        }
-        updateCarousel();
-    });
-
-    prevBtn.addEventListener('click', () => {
-        if (counter > 0) {
-            counter--;
-        }
-        updateCarousel();
-    });
-
-    // Update carousel on window resize and initial load
-    window.addEventListener('resize', updateCarousel);
-    updateCarousel();
-
-});
-
-</script>
-
-<script>
-     // This line uses PHP to securely insert the config from your new file
-    const firebaseConfig = <?php echo json_encode($firebaseConfig); ?>;
-
-    firebase.initializeApp(firebaseConfig);
-
-    // Find the logout link
-    const logoutLink = document.querySelector('a[href="includes/logout.php"]');
-    if (logoutLink) {
-        logoutLink.addEventListener('click', function(event) {
-            // Prevent the link from navigating immediately
-            event.preventDefault(); 
-            
-            firebase.auth().signOut().then(() => {
-                // Sign-out successful on the client-side. 
-                // Now, redirect to the PHP logout script to clear the server session.
-                window.location.href = 'includes/logout.php';
-            }).catch((error) => {
-                console.error('Sign out error', error);
-                // Even if there's an error, try to log out on the server
-                window.location.href = 'includes/logout.php';
-            });
-        });
-    }
-</script>
-
-<script>
-    // This code waits for the page content to be fully loaded before running
-    document.addEventListener('DOMContentLoaded', (event) => {
-        // Get the video element by its ID
-        const heroVideo = document.getElementById('hero-video');
-
-        // Set the playback speed
-        // 1.0 is normal speed
-        // 0.5 is half-speed
-        // 0.75 is 75% of normal speed
-        heroVideo.playbackRate = 0.75; 
-    });
-</script>
-
-<?php ob_end_flush(); // Flush the output buffer?>
-
-</body>
-</html>
+<?php require 'includes/footer.php'; ?>
